+++ src/pas/plugins/headers/utils.py
@@ -2,6 +2,8 @@
 from pas.plugins.headers.plugins import HeaderPlugin
 from Products.CMFCore.utils import getToolByName
 
+import six
+
 
 PLUGIN_ID = 'request_headers'
 
@@ -17,3 +19,19 @@ def get_plugin(context):
     if not isinstance(plugin, HeaderPlugin):
         return
     return plugin
+
+
+def safe_make_string(value):
+    """Make bytes/text value a string.
+
+    If value is a list/tuple, do this for each item.  Return a list.
+
+    If a value is an integer or some other non-bytes/string/text type,
+    let it remain the same.
+    """
+    if isinstance(value, (list, tuple, set)):
+        return [safe_make_string(v) for v in value]
+    try:
+        return six.ensure_str(value)
+    except TypeError:
+        return value